+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `isDisplayingAddressInformation` to `FormElement`
 - `isCustomPdfEditable` to `PDFConfiguration`
 
 ## [11.0.1] - 2025-03-19